refactor: consolidate agent prompt builders into zero/agent-prompt module
Move DISALLOWED_TOOLS, buildAgentToolsPrompt, and formatAgentIdentityPrompt
into a unified zero/agent-prompt.ts with a single buildAgentPrompt entry point.
Move buildIntegrationContext into zero/integration-context.ts to co-locate
with the zero layer. Delete the now-empty src/lib/integration-context.ts and
src/lib/agent-identity.ts.
