+++ turbo/apps/web/src/lib/email/handlers/inbound-trigger.ts
@@ -11,7 +11,7 @@ import {
   type HandlerResult,
 } from "./shared";
 import { createZeroRun } from "../../zero/zero-run-service";
-import { buildIntegrationContext } from "../../integration-context";
+import { buildIntegrationContext } from "../../zero/integration-context";
 import { generateCallbackSecret, getApiUrl } from "../../callback";
 import { getUserIdByEmail } from "../../auth/get-user-id-by-email";
 import { getOrgBySlug } from "../../org/org-cache-service";
