+++ turbo/apps/web/src/lib/zero/integration-context.ts
@@ -1,0 +1,35 @@
+/**
+ * The platform (integration) an agent run originates from.
+ * Used to inject context so the agent knows which channel it is operating in.
+ */
+type IntegrationPlatform = "Email" | "GitHub" | "Slack" | "Telegram";
+
+/**
+ * Build the integration context header prepended to agent run prompts.
+ */
+export function buildIntegrationContext(
+  platform: IntegrationPlatform,
+  options?: {
+    botUserId?: string;
+    channelId?: string;
+    channelType?: "channel" | "dm" | "group_dm";
+  },
+): string {
+  let context = `# Current Integration\nYou are currently running inside: ${platform}`;
+  if (options?.botUserId) {
+    context += `\nYour bot user ID: ${options.botUserId}`;
+  }
+  if (options?.channelId) {
+    context += `\nChannel ID: ${options.channelId}`;
+  }
+  if (options?.channelType) {
+    const typeLabel =
+      options.channelType === "dm"
+        ? "Direct message"
+        : options.channelType === "group_dm"
+          ? "Group direct message"
+          : "Channel";
+    context += `\nChannel type: ${typeLabel}`;
+  }
+  return context;
+}
